requirements.txt
faster-whisper
yt-dlp
ollama
